new_naev_xml.py -> naev_xml.py
