+++ utils/naev_xml.py
@@ -1,7 +1,12 @@
+# for testing:
 #!/usr/bin/env python3
 
 """
-A slim layer on top of xmltodict. Used as follows:
+An elementTree-based implementation of xmltodict. 10% slower than pure elementTree,
+but 2x faster than xmltodict with more functionality and correct output
+(empty element folding, new lines, no useless xml 1.0 header, etc.).
+
+Used as follows:
   o = naev_xml('input.xml')
   # Here o is just the regular xmltodict-generated dictionary, except:
 
@@ -39,11 +44,9 @@ A slim layer on top of xmltodict. Used as follows:
   # tags. This has been fixed.
 """
 
-from xmltodict import *
+import xml.etree.ElementTree as ET
 from sys import stderr
 from os import devnull
-import re
-
 
 def _numify( s ):
    try:
@@ -52,10 +55,11 @@ def _numify( s ):
    except:
       pass
 
-class _xml_node( dict ):
+class xml_node( dict ):
    def __init__ ( self, mapping, parent= None, key= None ):
-      mknode = lambda v: _xml_node(v) if isinstance(v, dict) else v
-      dict.__init__(self, {k: mknode(v) for k, v in mapping.items()})
+      mknode = lambda v: xml_node(v) if isinstance(v, dict) else v
+      self.attr = {k: v for k, v in mapping.items() if k[:1]=='@'}
+      dict.__init__(self, {k: v for k, v in mapping.items() if k[:1]!='@'})
       self._parent = parent
       self._key = key
 
@@ -65,11 +69,27 @@ class _xml_node( dict ):
       else:
          self._parent._change()
 
+   def parent( self ):
+      return self._parent, self._key
+
+   def __contains__ ( self, key ):
+      key = key[1:] if key[:1] == '$' else key
+      return key in self.attr if key[:1] == '@' else dict.__contains__(self, key)
+
+   def __getitem( self, key ):
+      return self.attr[key] if key[:1] == '@' else dict.__getitem__(self, key)
+
    def __getitem__( self, key ):
       if isinstance(key, str) and key[0] == '$':
-         return _numify(dict.__getitem__(self, key[1:]))
+         return _numify(self.__getitem(key[1:]))
       else:
-         return dict.__getitem__(self, key)
+         return self.__getitem(key)
+
+   def __setitem( self, key, val ):
+      if key[:1] == '@':
+         self.attr[key] = val
+      else:
+         dict.__setitem__(self, key, val)
 
    def __setitem__( self, key, val ):
       if isinstance(key, str) and key[0] == '$':
@@ -77,57 +97,131 @@ class _xml_node( dict ):
          if val is None:
             raise ValueError(str(val) + ' is not a number.')
          key = key[1:]
-      elif isinstance(val, dict) and not isinstance(val, _xml_node):
-         val = _xml_node(val, self, key)
+      elif isinstance(val, dict) and not isinstance(val, xml_node):
+         val = xml_node(val, self, key)
       dict.__setitem__(self, key, val)
       self._change()
 
    def __delitem__( self, key ):
-      dict.__delitem__(self, key[1:] if key[:1] == '$' else key)
+      if key[:1] == '@':
+         del self.attr[key]
+      else:
+         dict.__delitem__(self, key, val)
 
-   def parent( self ):
-      return self._parent, self._key
+   def __repr__( self ):
+      return dict.__repr__(self.attr | self)
 
-   def nodes( self, lookfor= None ):
-      for k, v in self.items():
-         if not lookfor or k in lookfor:
-            if _numify(v) is None:
-               yield self, k
-            else:
-               yield self, '$' + k
-         if isinstance(v, _xml_node):
-            for t in v.nodes(lookfor):
-               yield t
+   __str__ = __repr__
 
-class naev_xml( _xml_node ):
-   def __init__( self, fnam = devnull, read_only= False ):
+class _trusted_node( xml_node ):
+   def __init__ ( self, attr, parent= None, key= None ):
+      self.attr = {('@' + k): v for k, v in attr.items()}
+      dict.__init__(self, {})
+      self._parent = parent
+      self._key = key
+
+def _parse( node, par, key ):
+   d = _trusted_node(node.attrib, par, key)
+
+   for e in node:
+      if dict.__contains__(d, e.tag):
+         if not isinstance(d[e.tag], list):
+            dict.__setitem__(d, e.tag, [dict.__getitem__(d, e.tag)])
+         dict.__setitem__(d, e.tag, dict.__getitem__(d, e.tag) + [_parse(e, d, e.tag)])
+      else:
+         dict.__setitem__(d, e.tag, _parse(e, d, e.tag))
+
+   t = node.text
+   t = t and t.strip() or None
+   if d == {} and d.attr == {}:
+      return t
+   if t:
+      dict.__setitem__(d, '#text', t)
+   return d
+
+def _unparse_elt( v, k, indent):
+   out = ''
+   out += indent*' ' + '<' + k
+   if isinstance(v, dict):
+      for ka, va in v.attr.items():
+         out += ' ' + ka[1:] + '="' + va.replace('&', '&amp;') + '"'
+      if v == {}:
+         out += '/>\n'
+      else:
+         out += '>'
+         if not dict.__contains__(v, '#text'):
+            out += '\n'
+            out += unparse(v, indent+1)
+            out += indent*' '
+         else:
+            out += v['#text']
+         out += '</' + k + '>\n'
+   elif v:
+      out += '>'
+      out += str(v).replace('&', '&amp;')
+      out += '</' + k + '>\n'
+   else:
+      out += '/>\n'
+   return out
+
+def unparse( d , indent= 0 ):
+   out = ''
+   for k, v in d.items():
+      if isinstance(v, list):
+         for ve in v:
+            out += _unparse_elt(ve, k, indent)
+      else:
+         out += _unparse_elt(v, k, indent)
+
+   return out
+
+class naev_xml( xml_node ):
+   def __init__( self, fnam= devnull, read_only= False ):
       self._uptodate = True
       if type(fnam) != type('') or (not fnam.endswith('.xml') and fnam != devnull):
          raise Exception('Invalid xml filename ' + repr(fnam))
-      self._filename = devnull if read_only else fnam
       self.short = False
-      if fnam == devnull:
-         _xml_node.__init__(self, {})
-      else:
-         with open(fnam, 'r') as fp:
-            _xml_node.__init__(self, parse(fp.read(), process_comments= True))
+      _trusted_node.__init__(self, {}, None, None)
+      if fnam != devnull:
+         T = ET.parse(fnam).getroot()
+         dict.__setitem__(self, T.tag, _parse(T, self, T.tag))
+
+      self._filename = devnull if read_only else fnam
 
    def save( self ):
       if self._uptodate:
          stderr.write('Warning: saving unchanged file "' + self._filename + '".\n')
+
       with open(self._filename, 'w') as fp:
-         s = unparse(self, pretty= True, indent= ' ', full_document= False)
-         L = s.split('\n')
-         for i, l in enumerate(L):
-            if len(u := l.split('<#comment>',1)) == 2 and len(v := u[1].split('</#comment>',1)) == 2:
-               l = u[0] + '<!-- ' + v[0].replace('&lt;', '<').replace('&gt;', '>') + ' -->' + v[1]
-            l = re.sub(r'<([^ ]*)([^>]*)></\1>', r'<\1\2/>', l)
-            fp.write(l + '\n')
+         fp.write(unparse(self))
+
       self._uptodate = True
 
    def save_as( self, filename ):
       self._uptodate = self._uptodate and (filename == self._filename)
       self._filename = filename
+
+   def find( self, key, ref = False):
+      for d, k in self.nodes(lookfor = { key }):
+         return d if ref else d[k]
+
+   @staticmethod
+   def _nodes( node , lookfor ):
+      for k, v in node.items():
+         if not lookfor or k in lookfor:
+            if _numify(v) is None:
+               yield node, k
+            else:
+               yield node, '$' + k
+         if isinstance(v, dict):
+            for t in naev_xml._nodes(v, lookfor):
+               yield t
+
+   def nodes( self, lookfor= None ):
+      return self._nodes(self, lookfor)
+
+   def touch( self ):
+      self._uptodate = False
 
    def changed( self ):
       return not self._uptodate
@@ -135,7 +229,12 @@ class naev_xml( _xml_node ):
    def __del__( self ):
       if not self._uptodate and self._filename != devnull:
          stderr.write('Warning: unsaved file "' + self._filename + '" at exit.\n')
+"""
+if __name__ == '__main__':
+   from sys import argv
 
-   def find( self, key, ref = False):
-      for d, k in self.nodes(lookfor = { key }):
-         return d if ref else d[k]
+   for a in argv[1:]:
+      d = naev_xml(a)
+      d.touch()
+      d.save()
+"""
